Refactor Fact Modifiers: Required by Default, Optional (?). Closes #27 (#28)
## Summary

This PR refactors fact modifiers to change the default behaviour. Facts
are now required by default and always non-nullable.

## Changes

### New Behaviour

- **Facts are required by default** (breaking change)
- `?` operator marks a fact as optional
- **Facts are always non-nullable** - null values are not allowed for
facts

### Previous Behaviour

- Facts were optional by default
- `!` operator marked a fact as required
- No support for `?` operator
- No non-nullable validation

## Implementation Details

### AST Changes (`ast/fact.go`)

- Replaced `Required bool` field with:
- `Optional bool` - Whether the fact is optional (default: false, i.e.,
required)
- Removed nullable support - facts are always non-nullable
- Updated `NewFactStatement` signature: changed 5th parameter from
`required bool` to `optional bool` (inverted semantics)

### Parser Changes (`parser/fact.go`)

- Updated parser to handle only `?` operator (removed `!` operator
support)
- `?` sets `optional = true`
- Default behaviour: `optional = false` (required by default)
- Facts are always non-nullable

### Validation Changes (`index/policy.go`)

- Updated `AddFact` to check `!fact.Optional` instead of `fact.Required`
- Required facts (not optional) cannot have default values

### Runtime Changes (`runtime/executor.go`)

- Updated fact injection logic to:
  - Enforce required facts (error if not provided and not optional)
  - Always validate facts are not null (facts are always non-nullable)
  - Validate default values are not null

### Grammar Updates

- Updated `grammar.ebnf` and `grammar.peg` to reflect new fact
declaration syntax:
- `factDecl ::= 'fact' IDENT ('?'?) ':' typeRef ('as' IDENT)? ('default'
expr)?`

### Test Updates

- Updated all test files to use new `NewFactStatement` signature
(removed nullable parameter)
- Migrated old `required=false` (optional) → `optional=true`
- Migrated old `required=true` (required) → `optional=false`
- Added comprehensive parser tests for optional modifier (`?`)
- Added parser tests to verify `!` operator is rejected
- Added validation tests for required facts cannot have defaults
- Added validation tests for optional facts can have defaults

## Breaking Changes

1. **Facts are now required by default** - Previously optional facts
must now be marked with `?` to remain optional
2. **`!` operator removed** - No longer supported for facts
3. **Facts are always non-nullable** - Null values are not allowed
4. **API change** - `NewFactStatement` signature changed: parameter
changed from `required bool` to `optional bool` (inverted semantics)
5. **Required facts cannot have defaults** - Only optional facts can
have default values

## Migration Guide

For existing code:

- `fact user: User` → `fact user?: User` (if previously optional, add
`?`)
- `fact user!: User` → `fact user: User` (remove `!`, facts are always
non-nullable)
- `fact user: User default {...}` → `fact user?: User default {...}` (if
had default, add `?` to make optional)
- `fact user: User` → `fact user: User` (if previously required without
default, no change needed)

## Examples

```sentrie
-- Required fact (must be provided, always non-nullable)
fact user: User

-- Optional fact (can be omitted, but if provided, must be non-null)
fact context?: Context

-- Optional fact with default (can be omitted, uses default if not provided)
fact userId?: string default "anonymous"

-- Error: Required facts cannot have defaults
-- fact userId: string default "anonymous"  -- Indexing error!
```

+++ parser/expression_test.go
@@ -85,7 +85,7 @@ func (s *ExpressionTestSuite) TestParseExpressionIntegerLiteral() {
 
 	intLit, ok := expr.(*ast.IntegerLiteral)
 	s.True(ok)
-	s.Equal(int64(42), intLit.Value)
+	s.Equal(float64(42), intLit.Value)
 }
 
 // TestParseExpressionFloatLiteral tests parsing float literal expressions
